Changed amountofassets value type to decimal and all corresponding instances and functions of it

+++ CrytpWallet/Classes/Interfaces/IFungible.cs
@@ -9,7 +9,7 @@ namespace CrytpWallet.Classes.Interfaces
 {
     public  interface IFungible
     {
-        void GetFungible(FungibleAsset assetToAdd, int amount, bool newToken);
-        void SendFungible(FungibleAsset assetToRemove, int amount);
+        void GetFungible(FungibleAsset assetToAdd, decimal amount, bool newToken);
+        void SendFungible(FungibleAsset assetToRemove, decimal amount);
     }
 }
